Save progress on epsilon constrained method
I have no idea why a valid solver is generated, but it seems to ignore
the constraints relevant to the minimum return value.
When I try to limit the return (so that it doesn't exceed a threshold)
by changing the sign, it works.
I will have to examine this later, this commit is just a checkpoint as I
have to leave